Self-referentiality
When writing event listeners using jQuery, the this keyword plays an important role in the callback functions that run when an event happens. Inside a callback function on an event listener, this refers to event.currentTarget. So anywhere that you have been typing out event.currentTarget, you can replace it with the this keyword.