Event delegation
allows us to attach a single event listener, to a parent element, that will fire for all descendants matching a selector, whether those descendants exist now or are added in the future. his code says: when the page has loaded, be on the lookout for any clicks on the button (.click()) or an <li> element (.click()). If the user clicks the button, we append to the bottom of the unordered list. Specifically, we append a new <li> element with a random choice of text from the options cat, dog, and rock. Don't worry about understanding the details of the ['cat', 'dog', 'rock'][Math.floor(Math.random()*3)] line. You'll better understand how this works later in the course. For now, just know that it randomly chooses one of the strings between the brackets.