Event delegation another approach
we'll need a different approach: event delegation. Recall how event propagation works. When you click on the li, the event information will bubble up the DOM. Instead of watching for clicks on all <li> elements (regardless of when they were created), we'll watch for clicks on a parent element that we know existed at the time the page loaded. Since the event from the <li> will propagate to the parent, we can catch and handle it there.