Fizzbuzz game drill two
n this challenge, you'll reimplement FizzBuzz. This time you'll create an event listener that listens for when a user submits a form indicating how high to count. This event should cause the program to create the FizzBuzz sequence up to the number requested by the user.
For each element in the FizzBuzz sequence, your code should insert an element into the .js-results div that looks like this:
If the element's content is the word fizz, the div with .fizz-buzz-item should also get the class .fizz applied. If the element's content is the word buzz it should get the class .buzz applied. And if its content is the word fizzbuzz it should get the class fizzbuzz applied.